(parse) Consolidate Expression subclasses into Expression.ts
Those subclasses wound up being significantly smaller than I expected,
so keeping them all in one file actually works pretty well.
