--- src/parser/expressions/Literal.ts
@@ -1,11 +1,0 @@
-import { Expression } from "../Expression";
-import { Visitor } from "../Visitor";
-import { Literal as TokenLiteral } from "../../Token";
-
-export class Literal implements Expression {
-    constructor( readonly value: TokenLiteral) {}
-
-    accept <R extends Expression> (visitor: Visitor<R>): R {
-        return visitor.visitLiteral(this);
-    }
-}
