--- src/parser/expressions/index.ts
@@ -1,11 +1,0 @@
-export * from "./Assign";
-export * from "./Binary";
-export * from "./Call";
-export * from "./Get";
-export * from "./Grouping";
-export * from "./Literal";
-export * from "./Logical";
-export * from "./M";
-export * from "./Set";
-export * from "./Unary";
-export * from "./Variable";
